+++ sdk/dataprotection/arm-dataprotection/samples/v2/javascript/recoveryPointsListSample.js
@@ -16,7 +16,7 @@ require("dotenv").config();
  * This sample demonstrates how to Returns a list of Recovery Points for a DataSource in a vault.
  *
  * @summary Returns a list of Recovery Points for a DataSource in a vault.
- * x-ms-original-file: specification/dataprotection/resource-manager/Microsoft.DataProtection/stable/2023-11-01/examples/BackupInstanceOperations/ListRecoveryPoints.json
+ * x-ms-original-file: specification/dataprotection/resource-manager/Microsoft.DataProtection/stable/2024-04-01/examples/BackupInstanceOperations/ListRecoveryPoints.json
  */
 async function listRecoveryPointsInAVault() {
   const subscriptionId =