+++ sdk/dataprotection/arm-dataprotection/samples/v2/typescript/src/backupVaultsCreateOrUpdateSample.ts
@@ -10,7 +10,7 @@
 // Licensed under the MIT License.
 import {
   BackupVaultResource,
-  DataProtectionClient
+  DataProtectionClient,
 } from "@azure/arm-dataprotection";
 import { DefaultAzureCredential } from "@azure/identity";
 import * as dotenv from "dotenv";
@@ -21,7 +21,7 @@ dotenv.config();
  * This sample demonstrates how to Creates or updates a BackupVault resource belonging to a resource group.
  *
  * @summary Creates or updates a BackupVault resource belonging to a resource group.
- * x-ms-original-file: specification/dataprotection/resource-manager/Microsoft.DataProtection/stable/2023-11-01/examples/VaultCRUD/PutBackupVault.json
+ * x-ms-original-file: specification/dataprotection/resource-manager/Microsoft.DataProtection/stable/2024-04-01/examples/VaultCRUD/PutBackupVault.json
  */
 async function createBackupVault() {
   const subscriptionId =
@@ -36,23 +36,23 @@ async function createBackupVault() {
     properties: {
       featureSettings: { crossRegionRestoreSettings: { state: "Enabled" } },
       monitoringSettings: {
-        azureMonitorAlertSettings: { alertsForAllJobFailures: "Enabled" }
+        azureMonitorAlertSettings: { alertsForAllJobFailures: "Enabled" },
       },
       securitySettings: {
-        softDeleteSettings: { retentionDurationInDays: 14, state: "Enabled" }
+        softDeleteSettings: { retentionDurationInDays: 14, state: "Enabled" },
       },
       storageSettings: [
-        { type: "LocallyRedundant", datastoreType: "VaultStore" }
-      ]
+        { type: "LocallyRedundant", datastoreType: "VaultStore" },
+      ],
     },
-    tags: { key1: "val1" }
+    tags: { key1: "val1" },
   };
   const credential = new DefaultAzureCredential();
   const client = new DataProtectionClient(credential, subscriptionId);
   const result = await client.backupVaults.beginCreateOrUpdateAndWait(
     resourceGroupName,
     vaultName,
-    parameters
+    parameters,
   );
   console.log(result);
 }
@@ -61,7 +61,60 @@ async function createBackupVault() {
  * This sample demonstrates how to Creates or updates a BackupVault resource belonging to a resource group.
  *
  * @summary Creates or updates a BackupVault resource belonging to a resource group.
- * x-ms-original-file: specification/dataprotection/resource-manager/Microsoft.DataProtection/stable/2023-11-01/examples/VaultCRUD/PutBackupVaultWithMSI.json
+ * x-ms-original-file: specification/dataprotection/resource-manager/Microsoft.DataProtection/stable/2024-04-01/examples/VaultCRUD/PutBackupVaultWithCMK.json
+ */
+async function createBackupVaultWithCmk() {
+  const subscriptionId =
+    process.env["DATAPROTECTION_SUBSCRIPTION_ID"] ||
+    "0b352192-dcac-4cc7-992e-a96190ccc68c";
+  const resourceGroupName =
+    process.env["DATAPROTECTION_RESOURCE_GROUP"] || "SampleResourceGroup";
+  const vaultName = "swaggerExample";
+  const parameters: BackupVaultResource = {
+    identity: { type: "None" },
+    location: "WestUS",
+    properties: {
+      monitoringSettings: {
+        azureMonitorAlertSettings: { alertsForAllJobFailures: "Enabled" },
+      },
+      securitySettings: {
+        encryptionSettings: {
+          infrastructureEncryption: "Enabled",
+          kekIdentity: {
+            identityId:
+              "/subscriptions/85bf5e8c-3084-4f42-add2-746ebb7e97b2/resourcegroups/defaultrg/providers/Microsoft.ManagedIdentity/userAssignedIdentities/examplemsi",
+            identityType: "UserAssigned",
+          },
+          keyVaultProperties: {
+            keyUri:
+              "https://cmk2xkv.vault.azure.net/keys/Key1/0767b348bb1a4c07baa6c4ec0055d2b3",
+          },
+          state: "Enabled",
+        },
+        immutabilitySettings: { state: "Disabled" },
+        softDeleteSettings: { retentionDurationInDays: 0, state: "Off" },
+      },
+      storageSettings: [
+        { type: "LocallyRedundant", datastoreType: "VaultStore" },
+      ],
+    },
+    tags: { key1: "val1" },
+  };
+  const credential = new DefaultAzureCredential();
+  const client = new DataProtectionClient(credential, subscriptionId);
+  const result = await client.backupVaults.beginCreateOrUpdateAndWait(
+    resourceGroupName,
+    vaultName,
+    parameters,
+  );
+  console.log(result);
+}
+
+/**
+ * This sample demonstrates how to Creates or updates a BackupVault resource belonging to a resource group.
+ *
+ * @summary Creates or updates a BackupVault resource belonging to a resource group.
+ * x-ms-original-file: specification/dataprotection/resource-manager/Microsoft.DataProtection/stable/2024-04-01/examples/VaultCRUD/PutBackupVaultWithMSI.json
  */
 async function createBackupVaultWithMsi() {
   const subscriptionId =
@@ -76,29 +129,30 @@ async function createBackupVaultWithMsi() {
     properties: {
       featureSettings: { crossRegionRestoreSettings: { state: "Enabled" } },
       monitoringSettings: {
-        azureMonitorAlertSettings: { alertsForAllJobFailures: "Enabled" }
+        azureMonitorAlertSettings: { alertsForAllJobFailures: "Enabled" },
       },
       securitySettings: {
-        softDeleteSettings: { retentionDurationInDays: 14, state: "Enabled" }
+        softDeleteSettings: { retentionDurationInDays: 14, state: "Enabled" },
       },
       storageSettings: [
-        { type: "LocallyRedundant", datastoreType: "VaultStore" }
-      ]
+        { type: "LocallyRedundant", datastoreType: "VaultStore" },
+      ],
     },
-    tags: { key1: "val1" }
+    tags: { key1: "val1" },
   };
   const credential = new DefaultAzureCredential();
   const client = new DataProtectionClient(credential, subscriptionId);
   const result = await client.backupVaults.beginCreateOrUpdateAndWait(
     resourceGroupName,
     vaultName,
-    parameters
+    parameters,
   );
   console.log(result);
 }
 
 async function main() {
   createBackupVault();
+  createBackupVaultWithCmk();
   createBackupVaultWithMsi();
 }