+++ sdk/network/arm-network/samples-dev/networkWatchersGetTroubleshootingSample.ts
@@ -21,7 +21,7 @@ dotenv.config();
  * This sample demonstrates how to Initiate troubleshooting on a specified resource.
  *
  * @summary Initiate troubleshooting on a specified resource.
- * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2023-11-01/examples/NetworkWatcherTroubleshootGet.json
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2024-01-01/examples/NetworkWatcherTroubleshootGet.json
  */
 async function getTroubleshooting() {
   const subscriptionId = process.env["NETWORK_SUBSCRIPTION_ID"] || "subid";