+++ sdk/network/arm-network/samples/v33/javascript/firewallPolicyDeploymentsDeploySample.js
@@ -16,7 +16,7 @@ require("dotenv").config();
  * This sample demonstrates how to Deploys the firewall policy draft and child rule collection group drafts.
  *
  * @summary Deploys the firewall policy draft and child rule collection group drafts.
- * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2023-11-01/examples/FirewallPolicyDraftDeploy.json
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2024-01-01/examples/FirewallPolicyDraftDeploy.json
  */
 async function deployFirewallPolicyDraft() {
   const subscriptionId = process.env["NETWORK_SUBSCRIPTION_ID"] || "subid";