+++ sdk/network/arm-network/samples/v33/javascript/privateLinkServicesUpdatePrivateEndpointConnectionSample.js
@@ -16,7 +16,7 @@ require("dotenv").config();
  * This sample demonstrates how to Approve or reject private end point connection for a private link service in a subscription.
  *
  * @summary Approve or reject private end point connection for a private link service in a subscription.
- * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2023-11-01/examples/PrivateLinkServiceUpdatePrivateEndpointConnection.json
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2024-01-01/examples/PrivateLinkServiceUpdatePrivateEndpointConnection.json
  */
 async function approveOrRejectPrivateEndPointConnectionForAPrivateLinkService() {
   const subscriptionId = process.env["NETWORK_SUBSCRIPTION_ID"] || "subId";