+++ sdk/network/arm-network/samples/v33/javascript/webApplicationFirewallPoliciesCreateOrUpdateSample.js
@@ -16,7 +16,7 @@ require("dotenv").config();
  * This sample demonstrates how to Creates or update policy with specified rule set name within a resource group.
  *
  * @summary Creates or update policy with specified rule set name within a resource group.
- * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2023-11-01/examples/WafPolicyCreateOrUpdate.json
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2024-01-01/examples/WafPolicyCreateOrUpdate.json
  */
 async function createsOrUpdatesAWafPolicyWithinAResourceGroup() {
   const subscriptionId = process.env["NETWORK_SUBSCRIPTION_ID"] || "subid";