+++ sdk/network/arm-network/samples/v33/typescript/src/virtualHubsCreateOrUpdateSample.ts
@@ -18,7 +18,7 @@ dotenv.config();
  * This sample demonstrates how to Creates a VirtualHub resource if it doesn't exist else updates the existing VirtualHub.
  *
  * @summary Creates a VirtualHub resource if it doesn't exist else updates the existing VirtualHub.
- * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2023-11-01/examples/VirtualHubPut.json
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2024-01-01/examples/VirtualHubPut.json
  */
 async function virtualHubPut() {
   const subscriptionId = process.env["NETWORK_SUBSCRIPTION_ID"] || "subid";