+++ sdk/redis/arm-rediscache/samples/v8/javascript/accessPolicyAssignmentCreateUpdateSample.js
@@ -16,7 +16,7 @@ require("dotenv").config();
  * This sample demonstrates how to Adds the access policy assignment to the specified users
  *
  * @summary Adds the access policy assignment to the specified users
- * x-ms-original-file: specification/redis/resource-manager/Microsoft.Cache/stable/2023-08-01/examples/RedisCacheAccessPolicyAssignmentCreateUpdate.json
+ * x-ms-original-file: specification/redis/resource-manager/Microsoft.Cache/stable/2024-03-01/examples/RedisCacheAccessPolicyAssignmentCreateUpdate.json
  */
 async function redisCacheAccessPolicyAssignmentCreateUpdate() {
   const subscriptionId = process.env["REDIS_SUBSCRIPTION_ID"] || "subid";
@@ -34,7 +34,7 @@ async function redisCacheAccessPolicyAssignmentCreateUpdate() {
     resourceGroupName,
     cacheName,
     accessPolicyAssignmentName,
-    parameters
+    parameters,
   );
   console.log(result);
 }