+++ sdk/redis/arm-rediscache/samples/v8/javascript/redisRegenerateKeySample.js
@@ -16,7 +16,7 @@ require("dotenv").config();
  * This sample demonstrates how to Regenerate Redis cache's access keys. This operation requires write permission to the cache resource.
  *
  * @summary Regenerate Redis cache's access keys. This operation requires write permission to the cache resource.
- * x-ms-original-file: specification/redis/resource-manager/Microsoft.Cache/stable/2023-08-01/examples/RedisCacheRegenerateKey.json
+ * x-ms-original-file: specification/redis/resource-manager/Microsoft.Cache/stable/2024-03-01/examples/RedisCacheRegenerateKey.json
  */
 async function redisCacheRegenerateKey() {
   const subscriptionId = process.env["REDIS_SUBSCRIPTION_ID"] || "subid";