+++ sdk/search/search-documents/samples/v12/javascript/bufferedSenderAutoFlushSize.js
@@ -5,14 +5,14 @@
  * @summary Demonstrates the SearchIndexingBufferedSender with Autoflush based on size.
  */
 
+const { DefaultAzureCredential } = require("@azure/identity");
 const {
-  SearchIndexingBufferedSender,
-  AzureKeyCredential,
-  SearchClient,
   GeographyPoint,
+  SearchClient,
   SearchIndexClient,
+  SearchIndexingBufferedSender,
 } = require("@azure/search-documents");
-const { createIndex, documentKeyRetriever, WAIT_TIME, delay } = require("./setup");
+const { createIndex, delay, documentKeyRetriever, WAIT_TIME } = require("./setup");
 
 require("dotenv").config();
 
@@ -24,7 +24,6 @@ require("dotenv").config();
  * by default.
  */
 const endpoint = process.env.ENDPOINT || "";
-const apiKey = process.env.SEARCH_API_ADMIN_KEY || "";
 const TEST_INDEX_NAME = "example-index-sample-4";
 
 function getDocumentsArray(size) {
@@ -57,14 +56,14 @@ function getDocumentsArray(size) {
 }
 
 async function main() {
-  if (!endpoint || !apiKey) {
-    console.log("Make sure to set valid values for endpoint and apiKey with proper authorization.");
+  if (!endpoint) {
+    console.log("Be sure to set a valid endpoint with proper authorization.");
     return;
   }
 
   console.log(`Running SearchIndexingBufferedSender-uploadDocuments-With Auto Flush Sizes Sample`);
 
-  const credential = new AzureKeyCredential(apiKey);
+  const credential = new DefaultAzureCredential();
   const searchClient = new SearchClient(endpoint, TEST_INDEX_NAME, credential);
   const indexClient = new SearchIndexClient(endpoint, credential);
 
@@ -95,7 +94,7 @@ async function main() {
     });
 
     const documents = getDocumentsArray(1001);
-    bufferedClient.uploadDocuments(documents);
+    await bufferedClient.uploadDocuments(documents);
 
     await delay(WAIT_TIME);
 
@@ -112,7 +111,6 @@ async function main() {
   } finally {
     await indexClient.deleteIndex(TEST_INDEX_NAME);
   }
-  await delay(WAIT_TIME);
 }
 
 main();