+++ sdk/search/search-documents/samples/v12/typescript/src/bufferedSenderAutoFlushTimer.ts
@@ -5,16 +5,16 @@
  * @summary Demonstrates the SearchIndexingBufferedSender with Autoflush based on timer.
  */
 
+import { DefaultAzureCredential } from "@azure/identity";
 import {
-  SearchIndexingBufferedSender,
-  AzureKeyCredential,
-  SearchClient,
-  GeographyPoint,
-  SearchIndexClient,
   DEFAULT_FLUSH_WINDOW,
+  GeographyPoint,
+  SearchClient,
+  SearchIndexClient,
+  SearchIndexingBufferedSender,
 } from "@azure/search-documents";
-import { createIndex, documentKeyRetriever, WAIT_TIME, delay } from "./setup";
 import { Hotel } from "./interfaces";
+import { createIndex, delay, documentKeyRetriever, WAIT_TIME } from "./setup";
 
 import * as dotenv from "dotenv";
 dotenv.config();
@@ -27,22 +27,21 @@ dotenv.config();
  * by default.
  */
 const endpoint = process.env.ENDPOINT || "";
-const apiKey = process.env.SEARCH_API_ADMIN_KEY || "";
 const TEST_INDEX_NAME = "example-index-sample-5";
 
-export async function main() {
-  if (!endpoint || !apiKey) {
-    console.log("Make sure to set valid values for endpoint and apiKey with proper authorization.");
+export async function main(): Promise<void> {
+  if (!endpoint) {
+    console.log("Be sure to set a valid endpoint with proper authorization.");
     return;
   }
 
   console.log(`Running SearchIndexingBufferedSender-uploadDocuments-With Auto Flush Timer Sample`);
 
-  const credential = new AzureKeyCredential(apiKey);
+  const credential = new DefaultAzureCredential();
   const searchClient: SearchClient<Hotel> = new SearchClient<Hotel>(
     endpoint,
     TEST_INDEX_NAME,
-    credential
+    credential,
   );
   const indexClient: SearchIndexClient = new SearchIndexClient(endpoint, credential);
 
@@ -55,7 +54,7 @@ export async function main() {
       documentKeyRetriever,
       {
         autoFlush: true,
-      }
+      },
     );
 
     bufferedClient.on("batchAdded", (response: any) => {
@@ -76,7 +75,7 @@ export async function main() {
       console.log(response);
     });
 
-    bufferedClient.uploadDocuments([
+    await bufferedClient.uploadDocuments([
       {
         hotelId: "1",
         description:
@@ -112,7 +111,6 @@ export async function main() {
   } finally {
     await indexClient.deleteIndex(TEST_INDEX_NAME);
   }
-  await delay(WAIT_TIME);
 }
 
 main();